# flattenLet applied in bottomup traversal 