--- a/changelog/2023-11-06T14_40_14+01_00_fix_2598
+++ b/changelog/2023-11-06T14_40_14+01_00_fix_2598
@@ -1,0 +1,1 @@
+FIXED: Name duplication in generated Verilog involving reset synchronizer [#2598](https://github.com/clash-lang/clash-compiler/issues/2598)