@DigitalBrains1 2023-11-06
About 13% slower isn't really a minor effect though?